+++ providers/capv/upgrade/capv_test.go
@@ -16,6 +16,5 @@ var _ = Describe("Basic upgrade test", Ordered, func() {
 	cfg.AutoScalingSupported = false
 	// Disabled until https://github.com/giantswarm/roadmap/issues/1037
 	cfg.ExternalDnsSupported = false
-	cfg.GatewayAPISupported = false
 	common.Run(cfg)
 })