+++ providers/capz/standard/capz_test.go
@@ -17,8 +17,5 @@ var _ = Describe("Common tests", func() {
 	cfg := common.NewTestConfigWithDefaults()
 	// Disabled until https://github.com/giantswarm/roadmap/issues/2693
 	cfg.AutoScalingSupported = false
-	// Disabled until wildcard ingress support is added
-	cfg.ExternalDnsSupported = false
-	cfg.GatewayAPISupported = false
 	common.Run(cfg)
 })