+++ providers/capz/upgrade/capz_test.go
@@ -14,8 +14,5 @@ var _ = Describe("Basic upgrade test", Ordered, func() {
 	cfg := common.NewTestConfigWithDefaults()
 	// Disabled until https://github.com/giantswarm/roadmap/issues/2693
 	cfg.AutoScalingSupported = false
-	// Disabled until wildcard ingress support is added
-	cfg.ExternalDnsSupported = false
-	cfg.GatewayAPISupported = false
 	common.Run(cfg)
 })